test(ovh-shell): remove jest dependency
ref: DTCORE-3303
