+++ package.json
@@ -55,7 +55,6 @@
     "release": "scripts/release/release.sh",
     "start": "node scripts/start-application.mjs",
     "test": "turbo run test --concurrency=1",
-    "test:jest": "jest --runInBand",
     "test:scripts": "make test"
   },
   "resolutions": {
@@ -109,9 +108,6 @@
     "identity-obj-proxy": "^3.0.0",
     "inquirer": "^7.0.6",
     "inquirer-search-list": "1.2.6",
-    "jest": "29.7.0",
-    "jest-cucumber": "^3.0.1",
-    "jest-environment-jsdom": "^29.7.0",
     "lerna": "^5.5.0",
     "lint-staged": "^13.0.3",
     "lodash": "^4.17.11",
