+++ packages/manager/apps/container/package.json
@@ -68,7 +68,6 @@
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.2.1",
     "@types/dompurify": "^2.3.3",
-    "@types/jest": "^28.1.1",
     "@types/lodash-es": "^4.17.12",
     "@types/react": "^18.2.45",
     "@types/react-dom": "^18.0.0",
