+++ packages/manager/apps/dedicated/package.json
@@ -155,7 +155,8 @@
     "angular-mocks": "^1.7.x",
     "eslint-plugin-angular": "^4.0.1",
     "glob": "^7.1.4",
-    "jest": "^29.5.12",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "29.7.0",
     "webpack": "^5.76.2",
     "webpack-merge": "^5.0.2"
   },
